telegram/requirements.txt
python-telegram-bot[job-queue]>=20.6
httpx>=0.25
python-dotenv>=1.0